Revert "More improvements"
This reverts commit 982dbb235fe249f40a4bbbd4a5cc5a358645174c.

+++ java/l2server/gameserver/network/serverpackets/SystemMessage.java
@@ -71,12 +71,12 @@ public final class SystemMessage extends L2GameServerPacket
 
         public final int getIntValue()
         {
-            return (Integer) _value;
+            return ((Integer) _value).intValue();
         }
 
         public final long getLongValue()
         {
-            return (Long) _value;
+            return ((Long) _value).longValue();
         }
 
         public final int[] getIntArrayValue()
